moved logic for taking a screengrab into check_requirements file, renamed file to save_image.py
